requirements.txt
fastapi==0.111.0
uvicorn==0.30.5
pydantic==2.8.2
python-dotenv==1.0.1

httpx==0.27.0
PyYAML==6.0.2

SQLAlchemy==2.0.34
psycopg==3.2.1
alembic==1.13.2

passlib[bcrypt]==1.7.4
PyJWT==2.9.0
boto3==1.35.28

sse-starlette==2.1.0

redis==5.0.7
